# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk manajemen keuangan pribadi (Cashflow).",
        "title": "Cashflow API Gin",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bagas Rr",
            "url": "http://bagasrr.my.id"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/wallets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan semua dompet yang dimiliki pengguna, termasuk dompet pribadi dan grup.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Get All Wallets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.WalletResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/wallets/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan detail dompet berdasarkan ID, termasuk transaksi terkait.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Get Wallet By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.WalletResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "omitempty: sembunyikan jika nil"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.CategoryResponse": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Makanan"
                },
                "type": {
                    "type": "string",
                    "example": "EXPENSE"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "response.TransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 500
                },
                "category": {
                    "$ref": "#/definitions/response.CategoryResponse"
                },
                "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-01-31T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Gaji bulan Januari 2026"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "title": {
                    "type": "string",
                    "example": "Gaji Bulanan"
                },
                "user": {
                    "$ref": "#/definitions/response.UserResponse"
                }
            }
        },
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "user_role": {
                    "type": "string",
                    "example": "USER"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                },
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.WalletResponse"
                    }
                }
            }
        },
        "response.WalletResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 1000
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Tabungan"
                },
                "transaction_count": {
                    "type": "integer",
                    "example": 5
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.TransactionResponse"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}